refactor: update frontend dir structure
