+++ app/about/loading.tsx
@@ -1,4 +1,4 @@
-import AboutSkeleton from "@frontend/components/skeletons/AboutSkeleton";
+import AboutSkeleton from "@frontend/layouts/skeletons/AboutSkeleton";
 
 export default function Loading() {
   return <AboutSkeleton />;
