+++ app/projects/[slug]/page.tsx
@@ -1,12 +1,12 @@
 import { Metadata } from "next";
 import { useFetch } from "@backend/hooks/useFetch";
 import Section from "@frontend/components/common/Section";
-import GitHubBtn from "@frontend/layouts/GitHubBtn";
-import VisitBtn from "@frontend/components/buttons/VisitBtn";
-import BackBtn from "@frontend/layouts/BackBtn";
+import GitHubBtn from "@frontend/layouts/buttons/GitHubBtn";
+import VisitBtn from "@frontend/layouts/buttons/VisitBtn";
+import BackBtn from "@frontend/layouts/buttons/BackBtn";
 import Title from "@frontend/components/Title";
 import Description from "@frontend/components/Description";
-import Footer from "@frontend/components/footer/Footer";
+import Footer from "@frontend/layouts/footer/Footer";
 import Cldpicture from "@frontend/components/common/Cldpicture";
 
 type Props = { params: { slug: string } };
