+++ app/projects/loading.tsx
@@ -1,4 +1,4 @@
-import ProjectSkeleton from "@frontend/components/skeletons/ProjectSkeleton";
+import ProjectSkeleton from "@frontend/layouts/skeletons/ProjectSkeleton";
 
 export default function Loading() {
   return <ProjectSkeleton />;
