+++ app/projects/page.tsx
@@ -1,7 +1,7 @@
 import { Metadata } from "next";
-import ProjectSection from "@frontend/components/sections/ProjectSection";
-import BackToTop from "@frontend/components/buttons/BackToTop";
-import Footer from "@frontend/components/footer/Footer";
+import ProjectSection from "@frontend/layouts/sections/ProjectSection";
+import BackToTop from "@frontend/layouts/buttons/BackToTop";
+import Footer from "@frontend/layouts/footer/Footer";
 
 export const metadata: Metadata = {
   title: "Projects",
