+++ frontend/layouts/buttons/EmailBtn.tsx
