+++ frontend/layouts/buttons/EventBtn.tsx
@@ -1,5 +1,5 @@
 import { EventType } from "@backend/ts/shared-types/types";
-import { Button } from "./ui/Button";
+import { Button } from "../../components/ui/Button";
 
 export default function EventBtn({
   _id,
