+++ frontend/layouts/footer/Connect.tsx
