+++ frontend/layouts/header/Header.tsx
@@ -8,8 +8,8 @@ import Image from "next/image";
 import Link from "next/link";
 import tsukiwaLogoLight from "@assets/tsukiwa-logo.png";
 import tsukiwaLogoDark from "@assets/tsukiwa-logo-dark.png";
-import ThemeSwitchBtn from "@frontend/components/buttons/ThemeSwitchBtn";
-import Dropdown from "../Dropdown";
+import ThemeSwitchBtn from "@frontend/layouts/buttons/ThemeSwitchBtn";
+import Dropdown from "../../components/Dropdown";
 import Anchor from "./Anchor";
 import Version from "@frontend/components/Version";
 import useHeader from "@frontend/hooks/useHeader";
