+++ frontend/layouts/sections/AboutSection.tsx
