+++ frontend/layouts/sections/ProjectSection.tsx
