+++ frontend/layouts/sections/TechStackSection.tsx
