+++ frontend/layouts/sections/TopicSection.tsx
